+++ module/input/SensorFilter/data/config/webots/SensorFilter.yaml
@@ -41,7 +41,7 @@ ukf:
       rotation: [0.01, 0.01, 0.01, 0.01] # Not very sure of rotation (QX, QY, QZ, QW)
 
 # State = [roll, pitch, roll rate, pitch rate]'
-kalman_filter:
+kalman:
   # Continuous time process model: dx = Ac*x Essentially results in dx = [roll rate, pitch rate, 0, 0]'
   Ac:
     - [0, 0, 1, 0]
@@ -66,11 +66,11 @@ kalman_filter:
 
 mahony:
   # Mahony "Proportional" Gain on error
-  Kp: 10.8279
+  Kp: 1
   # Mahony "Integral" Gain used to compute bias
-  Ki: 17.8421
+  Ki: 0
   # Mahony Filter Initial bias
-  initial_bias: [0, 0, 0]
+  bias: [-0.0967263, 0.00640166, 0.197743]
   # Mahony Filter initial deadreckoned position and height
   initial_rTWw: [0, 0, 0.49]
